Add a program that finds and prints the first 98 Fibonacci numbers,
starting with 1 and 2, followed by a new line.
The numbers should be separated by comma, followed by a space ,
You are allowed to use the standard library
You are not allowed to use any other library (You can’t use GMP etc…)
You are not allowed to use long long, malloc, pointers, arrays/tables, or structures
You are not allowed to hard code any Fibonacci number (except for 1 and 2)

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 int main(void)
 {
 	int i;
-	unsigned long int n1 = 1;
-	unsigned long int n2 = 2;
-	unsigned long int nxt_num = n1 + n2;
+	double n1 = 1;
+	double n2 = 2;
+	double nxt_num = n1 + n2;
 
-	printf("%lu, %lu, ", n1, n2);
+	printf("%.0f, %.0f, ", n1, n2);
 	for (i = 3; i <= 98; i++)
 	{
-		printf("%lu", nxt_num);
+		printf("%.0f", nxt_num);
 		n1 = n2;
 		n2 = nxt_num;
 		nxt_num = n1 + n2;
